feat(order): make order table to update dynamically
Whenever the user click the plus (+) button next to each menu item an entry of
that item will be added to the table of orders. If more then a unit of that item
are ordered the exact number will be displayed next to the item name in
the table entry.

Total and per-item-entry prices are caculated and displayed automatically.